Various changes Pin to a specific toolchain version Format checking functionality Add a test to check the code formatting. Remove macro_use attribute
